se modulariza codigo
