+++ package.json
@@ -22,5 +22,8 @@
   "devDependencies": {
     "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2"
+  },
+  "dependencies": {
+    "date-fns": "^2.28.0"
   }
 }
