+++ src/app.js
@@ -1,245 +1,6 @@
-class UI {
-  static displayView(view, library) {
-    const tasks = library.getTasks(view);
-    UI.renderTasks(tasks);
-  }
-
-  static displayProjects(project, library) {
-    const projects = library.getTasksByProject(project);
-    UI.renderTasks(projects);
-  }
-
-  static renderTasks(tasks) {
-    const tasksTable = document.querySelector(".todo-list table tbody");
-    tasksTable.innerHTML = `
-        <tr>
-          <th class="task-name">Task</th>
-          <th class="task-due-date">Due Date</th>
-          <th class="task-priority">Priority</th>
-          <th class="">Status</th>
-        </tr>`;
-    tasks.forEach((task) => {
-      const row = `
-        <tr class="todo-task" data-id=${task.id}>
-          <td class="task-name">${task.title}</td>
-          <td class="task-due-date">${task.dueDate}</td>
-          <td class="task-priority">${task.priority}</td>
-          <td class="task-status">${task.status}</td>
-          <td class="task-status-change">
-            <span class="material-icons-outlined"> done </span>
-          </td>
-          <td class="task-delete">
-            <span class="material-icons-outlined"> delete </span>
-          </td>
-        </tr>`;
-      tasksTable.innerHTML += row;
-    });
-  }
-
-  static addProject(name) {
-    const projectList = document.querySelector(".projects-list");
-    let li = document.createElement("li");
-    li.innerText = name;
-    projectList.appendChild(li);
-  }
-
-  static removeProject() {}
-
-  static showForm(section) {
-    section.classList.remove("hidden");
-  }
-
-  static hideForm(section) {
-    section.classList.add("hidden");
-  }
-
-  static addTask(task) {
-    const tasksTable = document.querySelector(".todo-list table tbody");
-    const row = `
-        <tr class="todo-task" data-id=${task.id}>
-          <td class="task-name">${task.title}</td>
-          <td class="task-due-date">${task.dueDate}</td>
-          <td class="task-priority">${task.priority}</td>
-          <td class="task-status">${task.status}</td>
-          <td class="task-status-change">
-            <span class="material-icons-outlined"> done </span>
-          </td>
-          <td class="task-delete">
-            <span class="material-icons-outlined"> delete </span>
-          </td>
-        </tr>`;
-    tasksTable.innerHTML += row;
-  }
-
-  static removeTask(taskTable, task) {
-    taskTable.removeChild(task);
-  }
-
-  static displayTaskInfo() {}
-
-  static changeTaskStatus() {}
-
-  static sortTasks() {}
-}
-
-class Task {
-  constructor(title, description, priority, project, dueDate) {
-    this.title = title;
-    this.description = description;
-    this.priority = priority;
-    this.status = "Pending";
-    this.dueDate = dueDate;
-    if (project === "Home" || project === "Today" || project === "Week") {
-      this.project = "";
-    } else {
-      this.project = project;
-    }
-    this.id = Math.floor(Math.random() * 1000);
-  }
-
-  updateTaskStatus(status) {
-    this.status = status;
-  }
-
-  getDate() {
-    return this.dueDate;
-  }
-
-  setId(id) {
-    this.id = id;
-  }
-}
-
-class TaskLibrary {
-  constructor() {
-    this.tasks = [
-      {
-        title: "Task 1",
-        description: "Task 1 description",
-        priority: "High",
-        status: "Pending",
-        project: "Project-1",
-        dueDate: "2020-01-01",
-        id: 241,
-      },
-      {
-        title: "Task 2",
-        description: "Task 2 description",
-        priority: "Medium",
-        status: "Pending",
-        project: "Project-2",
-        dueDate: "2020-02-01",
-        id: 238,
-      },
-      {
-        title: "Task 3",
-        description: "Task 3 description",
-        priority: "Low",
-        status: "Pending",
-        project: "Project-3",
-        dueDate: "2022-04-20",
-        id: 237,
-      },
-      {
-        title: "Task 3",
-        description: "Task 3 description",
-        priority: "Low",
-        status: "Pending",
-        project: "",
-        dueDate: "2020-03-01",
-        id: 235,
-      },
-      {
-        title: "Task 3",
-        description: "Task 3 description",
-        priority: "Low",
-        status: "Pending",
-        project: "",
-        dueDate: "2020-03-01",
-        id: 234,
-      },
-      {
-        title: "Task 3",
-        description: "Task 3 description",
-        priority: "Low",
-        status: "Pending",
-        project: "Project-4",
-        dueDate: "2020-03-01",
-        id: 233,
-      },
-      {
-        title: "Task 3",
-        description: "Task 3 description",
-        priority: "Low",
-        status: "Pending",
-        project: "Project-4",
-        dueDate: "2020-03-01",
-        id: 231,
-      },
-      {
-        title: "Task 3",
-        description: "Task 3 description",
-        priority: "Low",
-        status: "Pending",
-        project: "Project-4",
-        dueDate: "2020-03-01",
-        id: 231,
-      },
-    ];
-    // if (localStorage.getItem("taskLibrary") === null) {
-    //   this.tasks = [];
-    //   localStorage.setItem("taskLibrary", JSON.stringify(this.tasks));
-    // } else {
-    //   this.tasks = JSON.parse(localStorage.getItem("taskLibrary"));
-    // }
-  }
-
-  addTask(task) {
-    // check if task id already exists
-    for (let i = 0; i < this.tasks.length; i++) {
-      if (task.id === this.tasks[i].id) {
-        let newId = Math.floor(Math.random() * 1000);
-        console.log(newId);
-        task.setId(newId);
-      }
-    }
-    console.log(task);
-    this.tasks.push(task);
-  }
-
-  getTasks(view) {
-    // returns a filtered/unfiltered array of tasks
-    switch (view) {
-      case "Home":
-        return this.tasks;
-        break;
-      case "Today":
-        return this.getTodayTasks();
-        break;
-      case "Week":
-        return this.getWeekTasks();
-        break;
-      default:
-        break;
-    }
-  }
-
-  getTodayTasks() {
-    return this.tasks.filter((task) => {
-      return task.dueDate === new Date().toISOString().slice(0, 10);
-    });
-  }
-
-  getWeekTasks() {
-    return this.tasks.filter((task) => {
-      return task.dueDate >= new Date().toISOString().slice(0, 10);
-    });
-  }
-
-  getTasksByProject(project) {
-    return this.tasks.filter((task) => task.project === project);
-  }
-}
+import UI from "./UI.js";
+import Task from "./Task.js";
+import TaskLibrary from "./TaskLibrary.js";
 
 let library = new TaskLibrary();
